// assembly/proto/system_call_ids.ts
export enum system_call_id {
  reserved_id = 0,
  prints = 1,
  verify_block_header = 2,
  apply_block = 3,
  apply_transaction = 4,
  apply_upload_contract_operation = 5,
  apply_call_contract_operation = 6,
  apply_set_system_call_operation = 7,
  put_object = 8,
  get_object = 9,
  get_next_object = 10,
  get_prev_object = 11,
  call_contract = 12,
  get_entry_point = 13,
  get_contract_arguments_size = 14,
  get_contract_arguments = 15,
  set_contract_result = 16,
  exit_contract = 17,
  get_head_info = 18,
  hash = 19,
  recover_public_key = 20,
  verify_block_signature = 21,
  verify_merkle_root = 22,
  get_transaction_payer = 23,
  get_transaction_rc_limit = 25,
  get_last_irreversible_block = 26,
  get_caller = 27,
  require_authority = 28,
  get_transaction_signature = 29,
  get_contract_id = 30,
  get_account_nonce = 31,
  get_account_rc = 32,
  consume_account_rc = 33,
  get_resource_limits = 34,
  consume_block_resources = 35,
}